pyproject.toml
[tool.poetry]
name = "inquisitor"
version = "0.1.0"
description = "Discord bot for tracking turn progress in Dominions 6"
authors = ["Kaylynn Morgan <mkaylynn7@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cattrs = "^23.2.3"
toml = "^0.10.2"
discord-py = "^2.3.2"
aiohttp = "^3.9.3"
attrs = "^23.2.0"
asyncpg = "^0.29.0"
yarl = "^1.9.4"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
